refactor to standard separate file per class in views
