+++ gbe/urls.py
@@ -29,8 +29,8 @@ from gbe.views import (
     MakeCostumeView,
     MakeVendorView,
     MakeSummerActView,
-    MergeBios,
-    MergeProfiles,
+    MergeProfileData,
+    MergeProfileExtra,
     MergeProfileSelect,
     PerformerUpdate,
     PersonaCreate,
@@ -277,10 +277,10 @@ urlpatterns = [
         MergeProfileSelect.as_view(),
         name='start_merge_users'),
     url(r'^profile/mergeprofile/(?P<pk>.*)/(?P<from_pk>.*)/?$',
-        MergeProfiles.as_view(),
+        MergeProfileData.as_view(),
         name='merge_profiles'),
     url(r'^profile/mergebio/(?P<pk>.*)/(?P<from_pk>.*)/?$',
-        MergeBios.as_view(),
+        MergeProfileExtra.as_view(),
         name='merge_bios'),
     url(r'^troupe/manage/?$',
         ReviewTroupesView.as_view(),
